up: pinctrl: Bring in legacy fpga and pinctrl driver
This commits just copies the legacy UP board platform driver and the
pinctrl/fpga one from ubilinux-3 branch.
The only change made was to obey to the following commit:
    58383c7 gpio: change member .dev to .parent

Kernel config option to built-in UP support added in Kconfig&Makefile.

UP1 board with FPGA1 protocol and old BIOS doesn't provide the necessary
information via ACPI to the kernel, therefore the new MFD FPGA driver
will not be loaded; instead this legacy platform driver will always be
loaded and based on the platform detected will load or not the pinctrl
one.