up: pinctrl: change SoC gpio references to use h/w pin offsets
Commit 03c4749dd6c7 "gpio / ACPI: Drop unnecessary ACPI GPIO to
 Linux GPIO translation" changed how the GPIO pins were numbered
by the CherryView pinctrl/gpio driver.  Consequently, the numeric
references to those GPIO pins need also to be updated in this driver
to match.  The numbering scheme is now based directly on the h/w
pin offsets within the controller (which are non-contiguous) instead
of the contiguous numbering that was previously used.

Signed-off-by: Dan O'Donovan <dan@emutex.com>